PBL3-61: Add Page Not Found 404

+++ src/main/java/com/laptrinhjavaweb/controller/admin/ProductManagementController.java
@@ -31,7 +31,7 @@ public class ProductManagementController {
 	@RequestMapping(value = "/san-pham/{slug_name}", method = RequestMethod.GET)
 	public String productManagementPage(Model model, @PathVariable String slug_name, HttpServletRequest request, HttpServletResponse respone, HttpSession session, @ModelAttribute("user") Users user) {
 		Users currentUser = (Users) session.getAttribute("LoginInfor");
-		if (currentUser.getIs_admin() == 1) {
+		if (currentUser != null && currentUser.getIs_admin() == 1) {
 			Departments departments = adminService.findBySlugName(slug_name);
 			List<Products> products = null;
 			if(departments.getSlug_name().equals("tat-ca-san-pham") == true) {
@@ -44,30 +44,30 @@ public class ProductManagementController {
 			model.addAttribute("departments", departments);
 			return "admin/productmanagement";
 		} else {
-			return null;
+			return "redirect:/khong-tim-thay-yeu-cau";
 		}
 	}
 	
 	@RequestMapping(value = "/san-pham/tao-moi/{slug_name}", method = RequestMethod.GET)
 	public String updateProductPage(@PathVariable String slug_name, Model model, HttpSession session) {
 		Users currentUser = (Users) session.getAttribute("LoginInfor");
-		if (currentUser.getIs_admin() == 1) {
+		if (currentUser != null && currentUser.getIs_admin() == 1) {
 			Departments departments = adminService.findBySlugName(slug_name);
 			model.addAttribute("departments",departments);
 			return "admin/createProducts";
 		} else {
-			return null;
+			return "redirect:/khong-tim-thay-yeu-cau";
 		}
 	}
 	
 	@RequestMapping(value = "/san-pham/delete/{slug_name}/{code}", method = RequestMethod.GET)
 	public String deleteOne(@PathVariable String code, @PathVariable String slug_name, HttpSession session) {
 		Users currentUser = (Users) session.getAttribute("LoginInfor");
-		if (currentUser.getIs_admin() == 1) {
+		if (currentUser != null && currentUser.getIs_admin() == 1) {
 			homeService.deleteOne(code);
 			return "redirect:/san-pham/"+slug_name; 
 		} else {
-			return null;
+			return "redirect:/khong-tim-thay-yeu-cau";
 		}
 	}
 	
@@ -81,14 +81,14 @@ public class ProductManagementController {
     @RequestMapping(value="/san-pham/{slug_name}/{code}/update", method = RequestMethod.GET)    
     public String findByCode(@PathVariable String code, Model m, HttpSession session){    
     	Users currentUser = (Users) session.getAttribute("LoginInfor");
-		if (currentUser.getIs_admin() == 1) {
+		if (currentUser != null && currentUser.getIs_admin() == 1) {
 			Products products= homeService.findByCode(code);
 	        Departments departments = adminService.findSlugNameByCode(code);
 	        m.addAttribute("products",products);  
 	        m.addAttribute("departments",departments);  
 	        return "admin/updateProduct";  
 		} else {
-			return null;
+			return "redirect:/khong-tim-thay-yeu-cau";
 		}
     }  
     
@@ -101,14 +101,14 @@ public class ProductManagementController {
     @RequestMapping(value="/san-pham/{slug_name}/{code}/mo-ta", method = RequestMethod.GET)    
     public String updateDescriptionProduct(@PathVariable String slug_name,@PathVariable String code, Model m, HttpSession session){    
     	Users currentUser = (Users) session.getAttribute("LoginInfor");
-		if (currentUser.getIs_admin() == 1) {
+		if (currentUser != null && currentUser.getIs_admin() == 1) {
 			DescriptionProduct desProduct= adminService.findProductDescription(code);
 	        Products products = homeService.findByCode(code);
 	        m.addAttribute("desProduct",desProduct);  
 	        m.addAttribute("products",products);
 	        return "admin/updateDescriptionProduct";  
 		} else {
-			return null;
+			return "redirect:/khong-tim-thay-yeu-cau";
 		}
     }
